feat: refactor SwapPanel for loading state and modularity
- Updated SwapPanel to accept loading, networkId, and tokenAddress as props for better state management.
- Introduced SwapPanelWrap component to encapsulate the swap logic and UI, enhancing modularity.
- Adjusted DesktopLayout and MobileLayout to pass loading state to SwapPanel, improving user experience during data fetching.
