+++ packages/kit/src/views/Market/MarketDetailV2/layouts/MobileLayout.tsx
@@ -57,7 +57,7 @@ const createOverviewPageComponent = () => {
 };
 
 export function MobileLayout() {
-  const { tokenAddress, networkId, tokenDetail } = useTokenDetail();
+  const { tokenAddress, networkId, tokenDetail, isLoading } = useTokenDetail();
 
   // Memoize Chart and Overview components to avoid re-creation on each render
   const ChartPageComponent = useMemo(
@@ -88,7 +88,11 @@ export function MobileLayout() {
       <Tab data={tabs} />
 
       {/* Swap panel placed outside the tabs for global visibility */}
-      <SwapPanel />
+      <SwapPanel
+        loading={isLoading}
+        networkId={networkId}
+        tokenAddress={tokenDetail?.address}
+      />
     </>
   );
 }
